refactor(Sender): expose to external
