+++ senders/console_sender.go
